air-quality-sensor/requirements.txt
Flask
setuptools
pyModbusTCP
paho-mqtt